[Bug 1776] sntp mishandles -t/--timeout and -a/--authentication.
Default to silent make rules, override with make V=1 or ./configure
--disable-silent-rules.

Correct --with-openssl-incdir defaulting with pkg-config.Correct
./build on systems without gtest available.

+++ ntpdc/ntpdc.c
@@ -1527,7 +1527,8 @@ help(
 
 		for (row = 0; row < rows; row++) {
 			for (word = row; word < words; word += rows)
-				fprintf(fp, "%-*.*s", col, col-1, list[word]);
+				fprintf(fp, "%-*.*s", (int)col,
+					(int)col - 1, list[word]);
 			fprintf(fp, "\n");
 		}
 	} else {
